Hack in branch merge heuristic specific for FreeBSD
We never want to accidentally "merge" from stable/ or releng/ into
head/. Only projects and possibly user should be recorded as merges, the
rest are cherrypicks.

TODO: push this into the ruleset, maybe only do it for projects/, do
some analysis on how many of those merge commits we would actually have
in the repo.